requirements.txt
asgiref==3.8.1
boto3==1.28.85
botocore==1.31.85
certifi==2024.12.14
charset-normalizer==3.4.0
Django==4.2.17
django-cors-headers==4.6.0
django-storages==1.13.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMuPDF==1.25.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.7.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.0.7
drf-yasg==1.21.8
